requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
requests
python-dotenvcommand: